Use cellexecutor to update runCell
Because runCell was stuck in the Private namespace, we had to copy a lot of code related to run actions. With the introduction of cellexecutors, we can just modify runCell directly. We do need to use the model only for updates now meaning the widget hack on the outputarea had to be updated.
