+++ package.json
@@ -96,7 +96,8 @@
         "schemaDir": "packages/dfnotebook-extension/schema",
         "outputDir": "dfnotebook/labextension",
         "disabledExtensions": [
-            "@jupyterlab/notebook-extension:tracker"
+            "@jupyterlab/notebook-extension:tracker",
+            "@jupyterlab/notebook-extension:cell-executor"
         ]
     },
     "styleModule": "style/index.js",
