+++ packages/dfnotebook-extension/package.json
@@ -135,7 +135,8 @@
     "schemaDir": "schema",
     "outputDir": "../../dfnotebook/labextension",
     "disabledExtensions": [
-      "@jupyterlab/notebook-extension:tracker"
+      "@jupyterlab/notebook-extension:tracker",
+      "@jupyterlab/notebook-extension:cell-executor"
     ]
   },
   "styleModule": "style/index.js",
