+++ packages/dfnotebook-extension/src/index.ts
@@ -39,6 +39,7 @@ import { IMainMenu } from '@jupyterlab/mainmenu';
 import * as nbformat from '@jupyterlab/nbformat';
 import {
   ExecutionIndicator,
+  INotebookCellExecutor,
   INotebookTracker,
   INotebookWidgetFactory,
   Notebook,
@@ -48,7 +49,8 @@ import {
   NotebookTracker,
   NotebookWidgetFactory,
   StaticNotebook,
-  ToolbarItems
+  ToolbarItems,
+  setCellExecutor
 } from '@jupyterlab/notebook';
 import { IObservableList } from '@jupyterlab/observables';
 import { ISettingRegistry } from '@jupyterlab/settingregistry';
@@ -82,8 +84,6 @@ import { DisposableSet } from '@lumino/disposable';
 import { Panel } from '@lumino/widgets';
 
 import {
-  DataflowNotebook,
-  DataflowNotebookActions,
   DataflowNotebookModel,
   DataflowNotebookModelFactory,
   DataflowNotebookPanel,
@@ -93,6 +93,8 @@ import {
 import { IRenderMimeRegistry } from '@jupyterlab/rendermime';
 import { IChangedArgs, PageConfig } from '@jupyterlab/coreutils';
 import { DataflowInputArea } from '@dfnotebook/dfcells';
+
+import { cellExecutor } from './cellexecutor';
 
 /**
  * The command IDs used by the notebook plugin.
@@ -314,7 +316,8 @@ const trackerPlugin: JupyterFrontEndPlugin<INotebookTracker> = {
   requires: [
     INotebookWidgetFactory,
     IDataflowNotebookWidgetFactory,
-    IEditorExtensionRegistry
+    IEditorExtensionRegistry,
+    INotebookCellExecutor    
   ],
   optional: [
     ICommandPalette,
@@ -626,6 +629,7 @@ const MiniMap: JupyterFrontEndPlugin<void> = {
 
 
 const plugins: JupyterFrontEndPlugin<any>[] = [
+  cellExecutor,
   factory,
   widgetFactoryPlugin,
   trackerPlugin,
@@ -750,6 +754,7 @@ function activateNotebookHandler(
   dfFactory: DataflowNotebookWidgetFactory.IFactory,
   // dfModelFactory: DataflowNotebookModelFactory.IFactory,
   extensions: IEditorExtensionRegistry,
+  executor: INotebookCellExecutor,
   palette: ICommandPalette | null,
   defaultBrowser: IDefaultFileBrowser | null,
   launcher: ILauncher | null,
@@ -761,6 +766,8 @@ function activateNotebookHandler(
   translator_: ITranslator | null,
   formRegistry: IFormRendererRegistry | null
 ): INotebookTracker {
+  setCellExecutor(executor);
+
   const translator = translator_ ?? nullTranslator;
   const sessionDialogs =
     sessionDialogs_ ?? new SessionContextDialogs({ translator });
@@ -1305,23 +1312,12 @@ function addCommands(
 
       if (current) {
         const { context, content } = current;
-        // !!! DATAFLOW NOTEBOOK UPDATE !!!
-        if (content instanceof DataflowNotebook) {
-          return DataflowNotebookActions.runAndAdvance(
-            content,
-            context.sessionContext,
-            sessionDialogs,
-            translator
-          );
-        } else {
-          return NotebookActions.runAndAdvance(
-            content,
-            context.sessionContext,
-            sessionDialogs,
-            translator
-          );
-        }
-        // !!! END DATAFLOW NOTEBOOK UPDATE !!!
+        return NotebookActions.runAndAdvance(
+          content,
+          context.sessionContext,
+          sessionDialogs,
+          translator
+        );
       }
     },
     isEnabled: args => (args.toolbar ? true : isEnabled()),
@@ -1341,23 +1337,12 @@ function addCommands(
 
       if (current) {
         const { context, content } = current;
-        // !!! DATAFLOW NOTEBOOK UPDATE !!!
-        if (content instanceof DataflowNotebook) {
-          return DataflowNotebookActions.run(
-            content,
-            context.sessionContext,
-            sessionDialogs,
-            translator
-          );
-        } else {
-          return NotebookActions.run(
-            content,
-            context.sessionContext,
-            sessionDialogs,
-            translator
-          );
-        }
-        // !!! END DATAFLOW NOTEBOOK UPDATE !!!
+        return NotebookActions.run(
+          content,
+          context.sessionContext,
+          sessionDialogs,
+          translator
+        );
       }
     },
     isEnabled
@@ -1376,23 +1361,12 @@ function addCommands(
 
       if (current) {
         const { context, content } = current;
-        // !!! DATAFLOW NOTEBOOK UPDATE !!!
-        if (content instanceof DataflowNotebook) {
-          return DataflowNotebookActions.runAndInsert(
-            content,
-            context.sessionContext,
-            sessionDialogs,
-            translator
-          );
-        } else {
-          return NotebookActions.runAndInsert(
-            content,
-            context.sessionContext,
-            sessionDialogs,
-            translator
-          );
-        }
-        // !!! END DATAFLOW NOTEBOOK UPDATE !!!
+        return NotebookActions.runAndInsert(
+          content,
+          context.sessionContext,
+          sessionDialogs,
+          translator
+        );
       }
     },
     isEnabled
@@ -1405,23 +1379,12 @@ function addCommands(
 
       if (current) {
         const { context, content } = current;
-        // !!! DATAFLOW NOTEBOOK UPDATE !!!
-        if (content instanceof DataflowNotebook) {
-          return DataflowNotebookActions.runAll(
-            content,
-            context.sessionContext,
-            sessionDialogs,
-            translator
-          );
-        } else {
-          return NotebookActions.runAll(
-            content,
-            context.sessionContext,
-            sessionDialogs,
-            translator
-          );
-        }
-        // !!! END DATAFLOW NOTEBOOK UPDATE !!!
+        return NotebookActions.runAll(
+          content,
+          context.sessionContext,
+          sessionDialogs,
+          translator
+        );
       }
     },
     isEnabled
@@ -1433,23 +1396,12 @@ function addCommands(
 
       if (current) {
         const { context, content } = current;
-        // !!! DATAFLOW NOTEBOOK UPDATE !!!
-        if (content instanceof DataflowNotebook) {
-          return DataflowNotebookActions.runAllAbove(
-            content,
-            context.sessionContext,
-            sessionDialogs,
-            translator
-          );
-        } else {
-          return NotebookActions.runAllAbove(
-            content,
-            context.sessionContext,
-            sessionDialogs,
-            translator
-          );
-        }
-        // !!! END DATAFLOW NOTEBOOK UPDATE !!!
+        return NotebookActions.runAllAbove(
+          content,
+          context.sessionContext,
+          sessionDialogs,
+          translator
+        );
       }
     },
     isEnabled: () => {
@@ -1468,23 +1420,12 @@ function addCommands(
 
       if (current) {
         const { context, content } = current;
-        // !!! DATAFLOW NOTEBOOK UPDATE !!!
-        if (content instanceof DataflowNotebook) {
-          return DataflowNotebookActions.runAllBelow(
-            content,
-            context.sessionContext,
-            sessionDialogs,
-            translator
-          );
-        } else {
-          return NotebookActions.runAllBelow(
-            content,
-            context.sessionContext,
-            sessionDialogs,
-            translator
-          );
-        }
-        // !!! END DATAFLOW NOTEBOOK UPDATE !!!
+        return NotebookActions.runAllBelow(
+          content,
+          context.sessionContext,
+          sessionDialogs,
+          translator
+        );
       }
     },
     isEnabled: () => {
@@ -1503,13 +1444,7 @@ function addCommands(
       const current = getCurrent(tracker, shell, args);
       if (current) {
         const { content } = current;
-        // !!! DATAFLOW NOTEBOOK UPDATE !!!
-        if (content instanceof DataflowNotebook) {
-          return DataflowNotebookActions.renderAllMarkdown(content);
-        } else {
-          return NotebookActions.renderAllMarkdown(content);
-        }
-        // !!! END DATAFLOW NOTEBOOK UPDATE !!!
+        return NotebookActions.renderAllMarkdown(content);
       }
     },
     isEnabled
@@ -2579,23 +2514,12 @@ function addCommands(
 
       current.content.extendContiguousSelectionTo(lastIndex);
 
-      // !!! DATAFLOW NOTEBOOK CHANGE !!!
-      if (current.content instanceof DataflowNotebook) {
-        void DataflowNotebookActions.run(
-          current.content,
-          current.sessionContext,
-          sessionDialogs,
-          translator
-        );
-      } else {
-        void NotebookActions.run(
-          current.content,
-          current.sessionContext,
-          sessionDialogs,
-          translator
-        );
-      }
-      // !!! END DATAFLOW NOTEBOOK CHANGE !!!
+      NotebookActions.run(
+        current.content,
+        current.sessionContext,
+        sessionDialogs,
+        translator
+      );
     }
   });
